Add stest package of manual, move some test case to this package
