+++ .travis.yml
@@ -3,7 +3,6 @@ language: java
 jdk: oraclejdk8
 addons:
   ssh_known_hosts:
-  - 47.93.9.236:22008
   - 47.94.231.67:22008
   - 47.94.10.122:22008
   sonarcloud:
