+++ src/test/java/stest/tron/wallet/assetissue/WalletTestAssetIssue002.java
@@ -66,7 +66,7 @@ public class WalletTestAssetIssue002 {
     Wallet.setAddressPreFixByte(CommonConstant.ADD_PRE_FIX_BYTE_MAINNET);
   }
 
-  @BeforeClass(enabled = false)
+  @BeforeClass(enabled = true)
   public void beforeClass() {
     channelFull = ManagedChannelBuilder.forTarget(fullnode)
         .usePlaintext(true)
@@ -98,10 +98,10 @@ public class WalletTestAssetIssue002 {
     }
   }
 
-  @Test(enabled = false)
+  @Test(enabled = true)
   public void testParticipateAssetissue() {
     try {
-      Thread.sleep(3000);
+      Thread.sleep(6000);
     } catch (InterruptedException e) {
       e.printStackTrace();
     }
@@ -134,7 +134,7 @@ public class WalletTestAssetIssue002 {
         toAddress, testKey003,blockingStubFull));
   }
 
-  @AfterClass(enabled = false)
+  @AfterClass(enabled = true)
   public void shutdown() throws InterruptedException {
     if (channelFull != null) {
       channelFull.shutdown().awaitTermination(5, TimeUnit.SECONDS);
