+++ src/test/java/stest/tron/wallet/assetissue/WalletTestAssetIssue011.java
@@ -71,7 +71,7 @@ public class WalletTestAssetIssue011 {
     Wallet.setAddressPreFixByte(CommonConstant.ADD_PRE_FIX_BYTE_MAINNET);
   }
 
-  @BeforeClass(enabled = false)
+  @BeforeClass(enabled = true)
   public void beforeClass() {
     PublicMethed.printAddress(testKeyForAssetIssue011);
     PublicMethed.printAddress(transferAssetCreateKey);
@@ -108,7 +108,7 @@ public class WalletTestAssetIssue011 {
     }
   }
 
-  @Test(enabled = false)
+  @Test(enabled = true)
   public void testTransferAssetCreateAccount() {
     //Transfer asset to create an account.
     Assert.assertTrue(PublicMethed
@@ -128,7 +128,7 @@ public class WalletTestAssetIssue011 {
   }
 
 
-  @AfterClass(enabled = false)
+  @AfterClass(enabled = true)
   public void shutdown() throws InterruptedException {
     if (channelFull != null) {
       channelFull.shutdown().awaitTermination(5, TimeUnit.SECONDS);
