+++ src/test/java/stest/tron/wallet/manual/WalletTestBlock001.java
@@ -1,4 +1,4 @@
-package stest.tron.wallet.block;
+package stest.tron.wallet.manual;
 
 import com.google.protobuf.ByteString;
 import io.grpc.ManagedChannel;
@@ -43,7 +43,7 @@ public class WalletTestBlock001 {
     Wallet.setAddressPreFixByte(CommonConstant.ADD_PRE_FIX_BYTE_MAINNET);
   }
 
-  @Test(enabled = false)
+  @Test(enabled = true)
   @BeforeClass
   public void beforeClass() {
     channelFull = ManagedChannelBuilder.forTarget(fullnode)
@@ -57,7 +57,7 @@ public class WalletTestBlock001 {
     blockingStubSolidity = WalletSolidityGrpc.newBlockingStub(channelSolidity);
   }
 
-  @Test(enabled = false)
+  @Test(enabled = true)
   public void testCurrentBlock() {
     Block currentBlock = blockingStubFull.getNowBlock(GrpcAPI.EmptyMessage.newBuilder().build());
     Assert.assertTrue(currentBlock.hasBlockHeader());
@@ -79,7 +79,7 @@ public class WalletTestBlock001 {
     newBlock.getTransactionsList();
   }
 
-  @Test(enabled = false)
+  @Test(enabled = true)
   public void testCurrentBlockFromSolidity() {
     Block currentBlock = blockingStubSolidity
         .getNowBlock(GrpcAPI.EmptyMessage.newBuilder().build());
